feat: connecting the table to client-js (#3)
In order to get `utop` working to run the program, this also reorganizes
all the code into `src`. Now `utop` started in bin can load `Dohickey`
and `Dohickey_unix` which is sufficient to start the server and run the
code. Items sent to the server are not encoded correctly, but I'm
merging this because the renamed files are getting away from github.

- [x] Give table.html its structure and a little bit of CSS
- [x] Split out some utils from Draw to help main work
- [x] Add event handlers in main to make the buttons work
- [x] Send events to the service worker